+++ src/webview/settings/SettingsTemplateGenerator.ts
@@ -16,6 +16,7 @@ import { getSettingsScript } from "./scripts/settingsManager";
 import { AnthropicSettings } from "./components/AnthropicSettings";
 import { CopilotSettings } from "./components/CopilotSettings";
 import { DeepSeekSettings } from "./components/DeepSeekSettings";
+import { GrokSettings } from "./components/GrokSettings";
 
 export class SettingsTemplateGenerator {
   private _settings: ExtensionSettings;
@@ -41,6 +42,7 @@ export class SettingsTemplateGenerator {
     const anthropicSettings = new AnthropicSettings(this._settings);
     const copilotSettings = new CopilotSettings(this._settings);
     const deepseekSettings = new DeepSeekSettings(this._settings);
+    const grokSettings = new GrokSettings(this._settings);
 
     return `
     <!DOCTYPE html>
@@ -69,6 +71,7 @@ export class SettingsTemplateGenerator {
         ${anthropicSettings.render()}
         ${copilotSettings.render()}
         ${deepseekSettings.render()}
+        ${grokSettings.render()}
         ${buttonGroup.render()}
       </div>
       ${getSettingsScript(this._settings, this._nonce)}